Add advanced dataset tab with galaxy challenge
